PMM-10667 move percona user reducer to own file
