+++ public/app/percona/shared/core/reducers/user.types.ts
@@ -1,0 +1,9 @@
+export interface UserDetails {
+  userId: number;
+  productTourCompleted: boolean;
+}
+
+export interface PerconaUserState extends UserDetails {
+  isAuthorized: boolean;
+  isPlatformUser: boolean;
+}
